% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{boot2pval}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{vector of sample 1}

\item{x2}{vector of sample 2}

\item{conf.level}{confidence level}

\item{iter}{number of iterations}

\item{mudiff}{difference in means being tested}

\item{test}{type of test}
}
\value{
pvalue and histogram
}
\description{
boot2pval
}
\examples{
x1=rnorm(100,0,1)
x2=rnorm(100,0,1)
boot2pval(x1,x2)
}
