% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Compute a confidence interval for the mean of a normal distribution}
\usage{
myci(data, alpha = 0.05)
}
\arguments{
\item{data}{a numeric vector of data}

\item{alpha}{the confidence level, default is 0.05}
}
\value{
a vector of length 2 with the lower and upper bounds of the confidence interval
}
\description{
Compute a confidence interval for the mean of a normal distribution
}
\examples{
myci(rnorm(100))
}
