% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb - my central limit theorem binomial}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability of success}
}
\value{
a histogram of the sample mean of a binomial distribution
}
\description{
mycltb - my central limit theorem binomial
}
\examples{
mycltb(10, 1000, 0.5)
}
