% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum likelihood estimation of beta parameters}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{sample vector}

\item{alpha}{vector of alpha values}

\item{beta}{vector of beta values}

\item{...}{additional arguments to be passed to \code{\link{contour}}}
}
\value{
a list with the following components:
\itemize{
\item{sample vector}
\item{coordinates of the maximum likelihood estimate}
\item{maximum likelihood}
\item{maximum likelihood estimate of alpha}
\item{maximum likelihood estimate of beta}
}
}
\description{
Maximum likelihood estimation of beta parameters
}
\examples{
x = rbeta(100, 2, 5)
alpha = seq(0.1, 10, 0.1)
beta = seq(0.1, 10, 0.1)
mymlbeta(x, alpha, beta)
}
