% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in the flight}

\item{gamma}{probability a plane will be truly overbooked}

\item{p}{probability of a "show"}
}
\value{
a named list containing nd, nc, N, p and gamma

plots of Objective function Vs n
}
\description{
calculates the number of tickets to be sold when the number of seats in the
flight is N and the probability of a "show" is p and gamma is the probability
a plane will be truly overbooked (more people show than there are seats)
 Use solely the appropriate discrete distribution
 Use the normal approximation
}
\details{
prints a named list containing nd, nc, N, p and gamma - where nd is
calculated using the discrete distribution and nc is the same calculated with
normal approximation.

creates a plot of Objective function Vs n, the objective function can be
constructed from the defining equation by simply making the equation equal
zero -- example 1-gamma-pnorm(...) = 0. The left-hand side can be called an
objective function. Make two of these -- one for the discrete and one for the
continuous case.
}
\examples{
ntickets(100,0.1,0.9)
ntickets(N=400,gamma = 0.02, p = 0.95)
ntickets(N=200,gamma = 0.02, p = 0.95)
}
