% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss_mss_rss_table.R
\name{tss_mss_rss_table}
\alias{tss_mss_rss_table}
\title{tss_mss_rss_table}
\usage{
tss_mss_rss_table(observed_values, predicted_values)
}
\arguments{
\item{observed_values}{A numeric vector of observed values.}

\item{predicted_values}{A numeric vector of predicted values.}
}
\value{
A data frame with the total sum of squares, model sum of squares, and residual sum of squares.
}
\description{
Compute the total sum of squares, model sum of squares, and residual sum of squares.
}
\examples{
observed_values <- c(1, 2, 3, 4, 5)
predicted_values <- c(1.5, 2.5, 3.5, 4.5, 5.5)
tss_mss_rss_table(observed_values, predicted_values)

}
