% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNegBin.R
\name{myNegBin}
\alias{myNegBin}
\title{myNegBin}
\usage{
myNegBin(y, r, p)
}
\arguments{
\item{y}{number of failures}

\item{r}{size}

\item{p}{probability}
}
\value{
the probability of y failures before r successes
}
\description{
Compute the probability of y failures before r successes
}
\examples{
myNegBin(5,2,0.5)
}
